requirements.txt
pandas
requests
plotly
seaborn
matplotlib
langchain
pydantic
openai
langchain-openai 
python-dotenv
streamlit
scikit-learn
langgraph
langsmith
langchain_anthropic
termcolor
docker
langchain_community
chromadb
bs4